// src/generated/models/email-template-model.ts
import type { AccessRole } from './access-role-model';
import type { Application } from './application-model';
import type { Environment } from './environment-model';

export interface EmailTemplate {
  /**
   * @displayName Email Template
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Email Template Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  emailTemplateName: string;
  /**
   * @displayName Access Role
   */
  accessRole?: Pick<AccessRole, 'id' | 'accessRoleName'>;
  /**
   * @displayName Active State
   * @validationRule Required for create/update operations
   */
  activeState: boolean;
  /**
   * @displayName Application
   * @validationRule Required for create/update operations
   */
  application: Pick<Application, 'id' | 'applicationName'>;
  /**
   * @displayName BCC Addresses
   */
  bCCAddresses?: string;
  /**
   * @displayName Body
   * @validationRule Required for create/update operations
   */
  body: string;
  /**
   * @displayName CC Addresses
   */
  cCAddresses?: string;
  /**
   * @displayName Environment
   */
  environment?: Pick<Environment, 'id' | 'environmentName'>;
  /**
   * @displayName Important
   * @validationRule Required for create/update operations
   */
  important: boolean;
  /**
   * @displayName Send From Address
   * @validationRule Required for create/update operations
   */
  sendFromAddress: string;
  /**
   * @displayName Subject
   * @validationRule Required for create/update operations
   */
  subject: string;
  /**
   * @displayName Support Contact
   * @validationRule Required for create/update operations
   */
  supportContact: string;
  /**
   * @displayName Version
   * @validationRule Required for create/update operations
   */
  version: number;
}

export const _EmailTemplate = 'EmailTemplate' as const;