// src/generated/models/guide-link-model.ts
import type { AccessRole } from './access-role-model';
import type { Application } from './application-model';
import type { Environment } from './environment-model';

export const GuideLinkTypeKeyToLabel = {
  'Guide': 'Guide',
  'ExternalLink': 'External Link',
  'DeveloperResource': 'Developer Resource'
} as const;
export type GuideLinkTypeKey = keyof typeof GuideLinkTypeKeyToLabel;

export interface GuideLink {
  /**
   * @displayName Guide Link
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Label
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  label: string;
  /**
   * @displayName Access Role
   */
  accessRole?: Pick<AccessRole, 'id' | 'accessRoleName'>;
  /**
   * @displayName Active State
   * @validationRule Required for create/update operations
   */
  activeState: boolean;
  /**
   * @displayName Application
   * @validationRule Required for create/update operations
   */
  application: Pick<Application, 'id' | 'applicationName'>;
  /**
   * @displayName Environment
   */
  environment?: Pick<Environment, 'id' | 'environmentName'>;
  /**
   * @displayName Type
   * @validationRule Required for create/update operations
   */
  typeKey: GuideLinkTypeKey;
  /**
   * @displayName URL
   * @validationRule Required for create/update operations
   */
  uRL: string;
}

export const _GuideLink = 'GuideLink' as const;